+++ back-end/routes/index.js
@@ -1,0 +1,7 @@
+const express = require('express');
+const router = express.Router();
+const apiRouter = require('./api');
+
+router.use('/api', apiRouter)
+
+module.exports = router;
